post-build: More deterministic output
So, I have a problem where we use prot-build in a project to generate a
crate that contains all our message definitions. Whenever people add
new messages they also need to run the `generate` tool. To ensure this,
we also run `generate` on CI and see if there is a diff.

Sadly, though, the output of prost-build is not deterministic: On my Mac
to order of messages in a file is not the same as when my coworker run
it. I'm not sure where this comes from exactly, but this my first test
to see if it's just some weird HashMap behavior.

- Use BTreeMap instead of HashMap for some more internal fields
- Explicitly sort messages and enums when generating code